requirements.txt
urllib3==1.26.15
tiktoken
fastapi>=0.104.1
uvicorn
python-dotenv
python-multipart
httpx
tenacity
fastapi-jwt
passlib
bcrypt
dashscope
sse_starlette
pydantic==2.2.1
tortoise-orm
aiomysql
loguru
ujson